Crypto.JOSE.Compact

Description

JWS, JWE and some related specifications provide for "compact" representations of certain types. This module defines classes and functions for working with such data.

Documentation

class FromCompact a where

Data that can be parsed from a compact representation.

Methods

fromCompact :: (AsError e, MonadError e m) => [ByteString] -> m a

decodeCompact :: (FromCompact a, AsError e, MonadError e m) => ByteString -> m a

Decode a compact representation.

class ToCompact a where

Data that can be converted to a compact representation.

Methods

toCompact :: a -> [ByteString]

encodeCompact :: ToCompact a => a -> ByteString

Encode data to a compact representation.