Crypto.JOSE.Error

Description

JOSE error types and helpers.

Running JOSE computations

runJOSE :: JOSE e m a -> m (Either e a)

Run the JOSE computation. Result is an Either e a where e is the error type (typically Error or JWTError)

unwrapJOSE :: JOSE e m a -> ExceptT e m a

Get the inner ExceptT value of the JOSE computation. Typically runJOSE would be preferred, unless you specifically need an ExceptT value.

newtype JOSE e m a

Constructors

JOSE (ExceptT e m a)

Base error type and class

data Error

All the errors that can occur.

Constructors

AlgorithmNotImplemented
A requested algorithm is not implemented

AlgorithmMismatch String
A requested algorithm cannot be used

KeyMismatch Text
Wrong type of key was given

KeySizeTooSmall
Key size is too small

OtherPrimesNotSupported
RSA private key with >2 primes not supported

RSAError Error
RSA encryption, decryption or signing error

CryptoError CryptoError
Various cryptonite library error cases

CompactDecodeError CompactDecodeError
Wrong number of parts in compact serialisation

JSONDecodeError String
JSON (Aeson) decoding error

NoUsableKeys
No usable keys were found in the key store

JWSCritUnprotected

JWSNoValidSignatures
AnyValidated policy active, and no valid signature encountered

JWSInvalidSignature
AllValidated policy active, and invalid signature encountered

JWSNoSignatures
AllValidated policy active, and there were no signatures on object that matched the allowed algorithms

class AsError r where

Minimal complete definition

_Error

Methods

_Error :: Prism' r Error

_AlgorithmNotImplemented :: Prism' r ()

_AlgorithmMismatch :: Prism' r String

_KeyMismatch :: Prism' r Text

_KeySizeTooSmall :: Prism' r ()

_OtherPrimesNotSupported :: Prism' r ()

_RSAError :: Prism' r Error

_CryptoError :: Prism' r CryptoError

_CompactDecodeError :: Prism' r CompactDecodeError

_JSONDecodeError :: Prism' r String

_NoUsableKeys :: Prism' r ()

_JWSCritUnprotected :: Prism' r ()

_JWSNoValidSignatures :: Prism' r ()

_JWSInvalidSignature :: Prism' r ()

_JWSNoSignatures :: Prism' r ()

JOSE compact serialisation errors

data InvalidNumberOfParts

The wrong number of parts were found when decoding a compact JOSE object.

Constructors

InvalidNumberOfParts Natural Natural
expected vs actual parts

expectedParts :: Getter InvalidNumberOfParts Natural

Get the expected or actual number of parts.

actualParts :: Getter InvalidNumberOfParts Natural

Get the expected or actual number of parts.

data CompactTextError

Bad UTF-8 data in a compact object, at the specified index

Constructors

CompactTextError Natural UnicodeException

data CompactDecodeError

An error when decoding a JOSE compact object. JSON decoding errors that occur during compact object processing throw JSONDecodeError.

Constructors

CompactInvalidNumberOfParts InvalidNumberOfParts

CompactInvalidText CompactTextError

_CompactInvalidNumberOfParts :: Prism' CompactDecodeError InvalidNumberOfParts

_CompactInvalidText :: Prism' CompactDecodeError CompactTextError