Crypto.JOSE.Types

Description

Data types for the JOSE library.

Documentation

newtype Base64Integer

A base64url encoded octet sequence interpreted as an integer.

The value is encoded in the minimum number of octets (no leading zeros) with the exception of 0 which is encoded as AA. A leading zero when decoding is an error.

Constructors

Base64Integer Integer

_Base64Integer :: Iso' Base64Integer Integer

data SizedBase64Integer

A base64url encoded octet sequence interpreted as an integer and where the number of octets carries explicit bit-length information.

Constructors

SizedBase64Integer Int Integer

makeSizedBase64Integer :: Integer -> SizedBase64Integer

Create a SizedBase64Integer' from an Integer.

checkSize :: Int -> SizedBase64Integer -> Parser SizedBase64Integer

Parsed a SizedBase64Integer with an expected number of bytes.

newtype Base64Octets

A base64url encoded octet sequence. Used for payloads, signatures, symmetric keys, salts, initialisation vectors, etc.

Constructors

Base64Octets ByteString

newtype Base64SHA1

A base64url encoded SHA-1 digest. Used for X.509 certificate thumbprints.

Constructors

Base64SHA1 ByteString

newtype Base64SHA256

A base64url encoded SHA-256 digest. Used for X.509 certificate thumbprints.

Constructors

Base64SHA256 ByteString

newtype Base64X509

A base64 encoded X.509 certificate.

Constructors

Base64X509 SignedCertificate

type SignedCertificate = SignedExact Certificate

A Signed Certificate

data URI

Represents a general universal resource identifier using its component parts.

For example, for the URI

  foo://anonymous@www.haskell.org:42/ghc?query#frag

the components are:

base64url :: (AsEmpty s1, AsEmpty s2, Cons s1 s1 Word8 Word8, Cons s2 s2 Word8 Word8) => Prism' s1 s2

Prism for encoding / decoding base64url.

To encode, review base64url. To decode, preview base64url.

Works with any combinations of strict/lazy ByteString.